+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2025_08_16_130000) do
+ActiveRecord::Schema[8.0].define(version: 2025_08_16_151323) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -163,7 +163,6 @@ ActiveRecord::Schema[8.0].define(version: 2025_08_16_130000) do
 
   create_table "users", force: :cascade do |t|
     t.string "username", null: false
-    t.string "name"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "email", default: "", null: false